[Add] Mode enum for controlling Instr parsing
When only producing C output, it is not necessary to perform some tasks,
e.g., fully resolving the `Orientation` of each component instance,
because the resulting information is *not* transferred to the C file.

This change adds an enumerated value to be used as a flag. At present
only 'normal' and 'minimal' modes are supported, but more could be added
for new purposes in the future.
When minimal is provided to the Instr visitor it passes the flag to the
constructor for component Instance which then skips setting (and
resolving) the Orientation of the component.

The minimal mode is selected when using the `mcstas-antlr` and
`mcxtrace-antlr` entrypoint scripts.
This should lead to faster Instr -> C conversion without loss of
information _in_ the C file.