[Add] Mode enum for controlling Instr parsing (#69)
* Add version to _POSIX_C_SOURCE define to allow use of `popen` in C code

* [Add] Mode enum for controlling Instr parsing

When only producing C output, it is not necessary to perform some tasks, e.g., fully resolving the `Orientation` of each component instance, because the resulting information is *not* transferred to the C file.

This change adds an enumerated value to be used as a flag. At present only 'normal' and 'minimal' modes are supported, but more could be added for new purposes in the future.
When minimal is provided to the Instr visitor it passes the flag to the constructor for component Instance which then skips setting (and resolving) the Orientation of the component.

The minimal mode is selected when using the `mcstas-antlr` and `mcxtrace-antlr` entrypoint scripts.
This should lead to faster Instr -> C conversion without loss of information _in_ the C file.

* [Ref] Stop rebuilding grammar files needlessly

The antlr4 output files only _need_ to be modified if their associated grammar (.g4) files have changed. Since seting up the java dependencies can have a significant time penalty (and can even time-out) stop checking for the need to regenerate at every import of `mccode-antlr`.

This allows the `antlr4-tools` dependency to become a developer-only requirement, and could also allow no longer distributing the grammar files inside of the Python module wheel.