[Fix] LocalRegistry known/fullname glob inconsistency
The `Reader` asks each `Registry` in turn if it 'knows' a filename.
If it does, that `Registry` is then used to retrieve the file 'fullname', 'path', or contents.
In the case of `LocalRegistry` the component is searched for in a more-restrictive way for the three latter cases, allowing for the possibility that a `LocalRegistry` 'knows' it has a file but can't find it.

This makes the later checks less stringent, which _could_ have the effect of allowing multiple matches for a file, especially if two copies of the same file (with modifications, presumably) exist in the same directory tree.